requirements.txt
asgiref==3.3.1
astroid==2.5.2
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==3.10.0
isort==5.8.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.2.0
pylint==2.7.4
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.4.1
